+++ nanoutils/__init__.py
@@ -4,12 +4,11 @@
 
 from .__version__ import __version__
 
-from . import typing_utils, empty, utils, numpy_utils, schema, set_attr, file_container
+from . import typing_utils, empty, utils, numpy_utils, schema, file_container
 from .typing_utils import *
 from .empty import *
 from .utils import *
 from .numpy_utils import *
-from .set_attr import *
 from .file_container import *
 from .schema import (
     Default, Formatter, supports_float, supports_int,
@@ -26,5 +25,5 @@ __all__ += empty.__all__
 __all__ += utils.__all__
 __all__ += numpy_utils.__all__
 __all__ += schema.__all__
-__all__ += set_attr.__all__
 __all__ += file_container.__all__
+__all__.sort()